use u64 as max terms type

+++ crates/sui-protocol-config/src/lib.rs
@@ -1147,7 +1147,7 @@ pub struct ProtocolConfig {
     group_ops_bls12381_uncompressed_g1_to_g1_cost: Option<u64>,
     group_ops_bls12381_uncompressed_g1_sum_base_cost: Option<u64>,
     group_ops_bls12381_uncompressed_g1_sum_cost_per_term: Option<u64>,
-    group_ops_bls12381_uncompressed_g1_sum_max_terms: Option<u32>,
+    group_ops_bls12381_uncompressed_g1_sum_max_terms: Option<u64>,
 
     // hmac::hmac_sha3_256
     hmac_hmac_sha3_256_cost_base: Option<u64>,
